% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnmp-generics.R
\name{predict.dnnmp}
\alias{predict.dnnmp}
\title{Function for prediction at new locations}
\usage{
\method{predict}{dnnmp}(
  object,
  nonref_coords,
  nonref_covars = NULL,
  probs = c(0.025, 0.5, 0.975),
  predict_sam = FALSE,
  nonref_par_labels = NULL,
  compute_control = NULL,
  verbose = TRUE,
  nreport = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class dnnmp.}

\item{nonref_coords}{an \eqn{m \times 2} matrix of the corresponding spatial coordinates.}

\item{nonref_covars}{an \eqn{m \times p} matrix of covariates for \eqn{m} new locations.}

\item{probs}{a vector of probabilities with values in \eqn{[0,1]} to generate the corresponding quantiles.}

\item{predict_sam}{logical; if true, posterior predictive samples for each new location are returned.}

\item{nonref_par_labels}{a vector of numeric values of length \eqn{m}, indicating which partition
each new location belongs to.}

\item{compute_control}{This argument is used to specify parameters for
numerically inverting cdf of the continuous-extended marginals for prediction.
Currently, the argument is a list that comprises: lb: lower bound of the support;
ub: upper bound of the support; maxit: maximum number of iterations;
tol: the desired accuracy (convergence tolerance). The default parameters 
are lb = -1, ub = maximum of the data + 100, maxit = 1000, and tol = 1e-5.
Parameters (cop_maxit and cop_tol) for generating samples from a conditional 
copula (e.g., Gumbel) using numerical methods are fixed now. 
Currently, cop_maxit = 1000 and cop_tol = 1e-5.}

\item{verbose}{logical; if true, progress of the prediction is printed to the screen.}

\item{nreport}{If \code{verbose = TRUE}, \code{n_report} defines the interval to report
the progress of prediction.}

\item{...}{additional arguments. No additional arguments are supported currently.}
}
\value{
The return object is a list that comprises:
}
\description{
This function produces posterior predictive samples for a set of new locations using a DNNMP.
}
\references{
Zheng, X., Kottas, A., and Sansó, B. (2023),
"Bayesian geostatistical modeling for discrete-valued processes".
Environmetrics, 34(7), e2805.
DOI: \href{https://doi.org/10.1002/env.2805}{https://doi.org/10.1002/env.2805}.

Zheng, X., Kottas, A., and Sansó, B (2023). 
"Nearest-neighbor mixture models for non-Gaussian spatial processes".
Bayesian Anal. 18 (4) 1191 - 1222, December 2023.
DOI: \href{https://doi.org/10.1214/23-BA1405}{https://doi.org/10.1214/23-BA1405}.
}
\author{
Xiaotian Zheng \email{xiaotianzheng42@gmail.com}
}
