% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnmp-generics.R
\name{rqr}
\alias{rqr}
\title{Randomized quantile residuals for model checking.}
\usage{
rqr(object)
}
\arguments{
\item{object}{an object of class dnnmp.}
}
\description{
This function produces posterior samples of the randomized quantile residuals for
reference locations. The function currently supports only the Poisson and negative binomial NNMPs.
}
\references{
Zheng, X., Kottas, A., and Sansó, B. (2023),
"Bayesian geostatistical modeling for discrete-valued processes".
Environmetrics, 34(7), e2805.
DOI: \href{https://doi.org/10.1002/env.2805}{https://doi.org/10.1002/env.2805}.

Zheng, X., Kottas, A., and Sansó, B (2023). 
"Nearest-neighbor mixture models for non-Gaussian spatial processes". 
Bayesian Anal. 18 (4) 1191 - 1222, December 2023.
DOI: \href{https://doi.org/10.1214/23-BA1405}{https://doi.org/10.1214/23-BA1405}.
}
\author{
Xiaotian Zheng \email{xiaotianzheng42@gmail.com}
}
