% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnmp-generics.R
\name{summary.dnnmp}
\alias{summary.dnnmp}
\title{Obtain posterior estimates of the DNNMP model parameters}
\usage{
\method{summary}{dnnmp}(object, pt = "mean", ci = c(0.025, 0.975), digit = 2, ...)
}
\arguments{
\item{object}{an object of class dnnmp.}

\item{pt}{a quoted keyword that specifies point estimate.
Supported keywords are \code{"mean"} and \code{"median"}.}

\item{ci}{a numeric vector of lower and upper probabilities with values in \eqn{[0,1]} 
to generate a credible interval for the unknown parameter(s).}

\item{digit}{an integer indicating the number of decimal places to be used.}

\item{...}{additional arguments. No additional arguments are supported currently.}
}
\value{
A vector of characters. 
        Each element of the vector is in the form: point estimate (lower percentile, upper percentile).
}
\description{
This function calculates point and interval estimates of unknown parameters
using their posterior samples
}
